Django_Blog/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
Pillow==8.3.1
pip==21.1.3
pytz==2021.1
setuptools==57.0.0
sqlparse==0.4.1
wheel==0.36.2